backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.10.0
httpx==0.24.1
whisper==20231117
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pydantic==2.5.3
websockets==12.0
